+++ .projenrc.ts
@@ -679,7 +679,9 @@ const tmpToolkitHelpers = configureProject(
     ],
     deps: [
       cloudAssemblySchema.name,
+      cloudFormationDiff,
       'archiver',
+      'chalk@4',
       'glob',
       'semver',
       'uuid',
