+++ packages/@aws-cdk/tmp-toolkit-helpers/.projen/tasks.json
@@ -77,7 +77,7 @@
       "name": "gather-versions",
       "steps": [
         {
-          "exec": "node -e \"require(require.resolve('cdklabs-projen-project-types/lib/yarn/gather-versions.exec.js')).cliMain()\" ",
+          "exec": "node -e \"require(require.resolve('cdklabs-projen-project-types/lib/yarn/gather-versions.exec.js')).cliMain()\" @aws-cdk/cloudformation-diff=major",
           "receiveArgs": true
         }
       ]
