+++ packages/@aws-cdk/tmp-toolkit-helpers/package.json
@@ -56,7 +56,9 @@
   },
   "dependencies": {
     "@aws-cdk/cloud-assembly-schema": "^0.0.0",
+    "@aws-cdk/cloudformation-diff": "^0.0.0",
     "archiver": "^7.0.1",
+    "chalk": "4",
     "glob": "^11.0.1",
     "semver": "^7.7.1",
     "uuid": "^11.1.0",
