+++ packages/@aws-cdk/tmp-toolkit-helpers/src/api/cloudformation/stack-helpers.ts
@@ -1,0 +1,11 @@
+export interface Template {
+  Parameters?: Record<string, TemplateParameter>;
+  [section: string]: any;
+}
+
+export interface TemplateParameter {
+  Type: string;
+  Default?: any;
+  Description?: string;
+  [key: string]: any;
+}
