+++ packages/@aws-cdk/tmp-toolkit-helpers/src/api/diff/index.ts
@@ -1,0 +1,1 @@
+export * from './diff';
