+++ packages/@aws-cdk/tmp-toolkit-helpers/src/api/index.ts
@@ -1,4 +1,6 @@
 export * from './cloud-assembly';
+export * from './cloudformation';
+export * from './diff';
 export * from './io';
 export * from './toolkit-error';
 export * from './require-approval';
