+++ packages/@aws-cdk/tmp-toolkit-helpers/tsconfig.dev.json
@@ -38,5 +38,9 @@
   "exclude": [
     "node_modules"
   ],
-  "references": []
+  "references": [
+    {
+      "path": "../cloudformation-diff"
+    }
+  ]
 }
