+++ packages/@aws-cdk/tmp-toolkit-helpers/tsconfig.json
@@ -36,5 +36,9 @@
     "src/**/*.ts"
   ],
   "exclude": [],
-  "references": []
+  "references": [
+    {
+      "path": "../cloudformation-diff"
+    }
+  ]
 }
